Separate code & others shits
