+++ src/routes/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import CurrentNormal from '../components/current/CurrentNormal.svelte';
 	import DailyNormal from '../components/daily/DailyNormal.svelte';
-	import Hourly from '../components/hourly/hourly.svelte';
+	import Hourly from '../components/hourly/Hourly.svelte';
 	import { onMount } from 'svelte';
 	import { getCurrentWeather } from '../lib/getData';
